+++ packages/client/src/descriptors.ts
@@ -47,6 +47,7 @@ export type ChainDefinition = {
     apis: ApisTypedef<any>
   }
   asset: PlainDescriptor<any>
+  extensions?: Record<string, { value?: any; additionalSigned?: any }>
   metadataTypes: Promise<Uint8Array>
   getMetadata: () => Promise<Uint8Array>
   genesis: HexString | undefined